Add demo FMR list view
Adds a basic list view demo to the FMR tab. also increases the size of the application. This list view can be added to later on. To add your own list view to the other tabs look at the changes on this commit to the project to see what methods you need to add.